# server/routes/generated/messages.generated.js
// Auto-generated route file for messages
// Generated on: 2025-07-24T15:59:20.352Z
// Environment: local

const express = require('express');
const router = express.Router();

// TODO: Import your models
// const Messages = require('../models/Messages');

// TODO: Import middleware if needed
// const { validateRequest } = require('../middleware/validation');

// Routes

// Get all messages
// Retrieve all messages from the database
router.get('/api/messages', async (req, res) => {
  try {
    // TODO: Implement get all logic
    // const results = await Model.find();
    res.status(200).json({ message: 'Get all - Implementation needed', data: [] });
  } catch (error) {
    console.error('Error in getMessages:', error);
    res.status(500).json({ error: 'Internal server error' });
  }
});

// Create a new message
// Add a new message to the database
router.post('/api/messages', async (req, res) => {
  try {
    // Request body
    const data = req.body;
    // TODO: Implement create logic
    // const newResource = new Model(data);
    // const saved = await newResource.save();
    res.status(201).json({ message: 'Create - Implementation needed', data: data });
  } catch (error) {
    console.error('Error in postMessages:', error);
    res.status(500).json({ error: 'Internal server error' });
  }
});

// Get message by ID
// Retrieve a specific message by its ID
router.get('/api/messages/:id', async (req, res) => {
  try {
    // Path parameters: id
    const id = req.params.id;
    // TODO: Implement get by ID logic
    // const result = await Model.findById(id);
    // if (!result) {
    //   return res.status(404).json({ error: 'Resource not found' });
    // }
    res.status(200).json({ message: 'Get by ID - Implementation needed', data: {} });
  } catch (error) {
    console.error('Error in getMessagesById:', error);
    res.status(500).json({ error: 'Internal server error' });
  }
});

// Delete message by ID
// Remove a specific message from the database
router.delete('/api/messages/:id', async (req, res) => {
  try {
    // Path parameters: id
    const id = req.params.id;
    // TODO: Implement delete logic
    // const deleted = await Model.findByIdAndDelete(id);
    // if (!deleted) {
    //   return res.status(404).json({ error: 'Resource not found' });
    // }
    res.status(200).json({ message: 'Delete - Implementation needed' });
  } catch (error) {
    console.error('Error in deleteMessagesById:', error);
    res.status(500).json({ error: 'Internal server error' });
  }
});

module.exports = router;