# server/routes/generated/users.generated.js
// Auto-generated route file for users
// Generated on: 2025-07-24T15:59:20.352Z
// Environment: local

const express = require('express');
const router = express.Router();

// TODO: Import your models
// const Users = require('../models/Users');

// TODO: Import middleware if needed
// const { validateRequest } = require('../middleware/validation');

// Routes

// Get all users
// Retrieve all users from the database
router.get('/api/users', async (req, res) => {
  try {
    // TODO: Implement get all logic
    // const results = await Model.find();
    res.status(200).json({ message: 'Get all - Implementation needed', data: [] });
  } catch (error) {
    console.error('Error in getUsers:', error);
    res.status(500).json({ error: 'Internal server error' });
  }
});

// Create a new user
// Add a new user to the database
router.post('/api/users', async (req, res) => {
  try {
    // Request body
    const data = req.body;
    // TODO: Implement create logic
    // const newResource = new Model(data);
    // const saved = await newResource.save();
    res.status(201).json({ message: 'Create - Implementation needed', data: data });
  } catch (error) {
    console.error('Error in postUsers:', error);
    res.status(500).json({ error: 'Internal server error' });
  }
});

// Get user by ID
// Retrieve a specific user by their ID
router.get('/api/users/:id', async (req, res) => {
  try {
    // Path parameters: id
    const id = req.params.id;
    // TODO: Implement get by ID logic
    // const result = await Model.findById(id);
    // if (!result) {
    //   return res.status(404).json({ error: 'Resource not found' });
    // }
    res.status(200).json({ message: 'Get by ID - Implementation needed', data: {} });
  } catch (error) {
    console.error('Error in getUsersById:', error);
    res.status(500).json({ error: 'Internal server error' });
  }
});

// Update user by ID
// Update a specific user's information
router.put('/api/users/:id', async (req, res) => {
  try {
    // Path parameters: id
    const id = req.params.id;
    // Request body
    const data = req.body;
    // TODO: Implement update logic
    // const updated = await Model.findByIdAndUpdate(id, data, { new: true });
    // if (!updated) {
    //   return res.status(404).json({ error: 'Resource not found' });
    // }
    res.status(200).json({ message: 'Update - Implementation needed', data: data });
  } catch (error) {
    console.error('Error in putUsersById:', error);
    res.status(500).json({ error: 'Internal server error' });
  }
});

// Delete user by ID
// Remove a specific user from the database
router.delete('/api/users/:id', async (req, res) => {
  try {
    // Path parameters: id
    const id = req.params.id;
    // TODO: Implement delete logic
    // const deleted = await Model.findByIdAndDelete(id);
    // if (!deleted) {
    //   return res.status(404).json({ error: 'Resource not found' });
    // }
    res.status(200).json({ message: 'Delete - Implementation needed' });
  } catch (error) {
    console.error('Error in deleteUsersById:', error);
    res.status(500).json({ error: 'Internal server error' });
  }
});

module.exports = router;